Blockmap consistancy in T_Pusher()

+++ Radix/p_spec.pas
@@ -3608,10 +3608,10 @@ begin
     tmbbox[BOXRIGHT]  := p.x + radius;
     tmbbox[BOXLEFT]   := p.x - radius;
 
-    xl := MapBlockIntX(int64(tmbbox[BOXLEFT]) - bmaporgx - MAXRADIUS);
-    xh := MapBlockIntX(int64(tmbbox[BOXRIGHT]) - bmaporgx + MAXRADIUS);
-    yl := MapBlockIntY(int64(tmbbox[BOXBOTTOM]) - bmaporgy - MAXRADIUS);
-    yh := MapBlockIntY(int64(tmbbox[BOXTOP]) - bmaporgy + MAXRADIUS);
+    xl := MapBlockIntX(int64(tmbbox[BOXLEFT]) - int64(bmaporgx) - MAXRADIUS);
+    xh := MapBlockIntX(int64(tmbbox[BOXRIGHT]) - int64(bmaporgx) + MAXRADIUS);
+    yl := MapBlockIntY(int64(tmbbox[BOXBOTTOM]) - int64(bmaporgy) - MAXRADIUS);
+    yh := MapBlockIntY(int64(tmbbox[BOXTOP]) - int64(bmaporgy) + MAXRADIUS);
 
     bx := xl;
     while bx <= xh do
